modified BuravlevTest
